Create ResolveDB_backup.sh
Bash script to be copied in "/usr/local/bin" (please edit the cronjob creator script accordingly if different). It actually dumps all the local system's PostgreSQL database to a file on a local path, then compresses it and moves both copies to a remote folder.
Default local folder is "/root"; default remote folder is "/mnt/NAS" and PostgreSQL binaries are expected to be in the Path environment variable.